Factor tests to reflect package factor
